remove arrayloaderbase entirely
